Refactoring from v3 left an incorrect lock_upgrade
If assign() is never called for some reason, mutex_ will remain locked. Downgrading to lock_shared is atomic, and attempting to downgrade if already downgraded should have no effect. See:

https://www.boost.org/doc/libs/1_68_0/doc/html/thread/synchronization.html

"Lock ownership acquired through a call to lock_upgrade() must be released through a call to unlock_upgrade(). If the ownership type is changed through a call to one of the unlock_xxx_and_lock_yyy() functions, ownership must be released through a call to the unlock function corresponding to the new level of ownership."