end_write() must return true unless flush() fails
my debugging indicates there is a condition in which another thread has deleted the flush_lock file already when end_write() is called, but that's not an error condition and should not cause callers to conclude that an error::store_lock_failure has occurred. all is well with the tables at this point and the only reason we have just tried to delete a non-existent file is that we are treating the flush locking mechanism as idempotent. Multiple calls to create the flush_lock file are not an error condition, either. if boost::filesystem::remove() does encounter an error, it won't signal that with a FALSE return value in any case because boost throws exceptions for that. see:

https://www.boost.org/doc/libs/1_46_1/libs/filesystem/v3/doc/reference.html#remove